+++ src/main/java/jp/crafterkina/SurviveInSociety/common/CommonProxy.java
@@ -3,8 +3,8 @@
  * All rights reserved.
  */
 
-package jp.crafterkina.SurviveInSociety.server;
+package jp.crafterkina.SurviveInSociety.common;
 
 import jp.crafterkina.KinaCore.misclib.base.fle.AbstractFMLStateEvent;
 
-public class ServerProxy extends AbstractFMLStateEvent{}
+public class CommonProxy extends AbstractFMLStateEvent{}
